second commit

+++ public/assets/js/listBeer.js
@@ -1,4 +1,5 @@
-function Beer(data) {
+function BeerModel(data) {
+    this.id = data.id;
     this.name = ko.observable(data.name);
     this.brewery = ko.observable(data.brewery);
     this.type = ko.observable(data.type);
@@ -14,7 +15,7 @@ function Beer(data) {
     this.image_url = ko.observable(data.image_url);
 }
 
-function BeerViewModel() {
+function BeerListViewModel() {
     const self = this;
 
     self.beers = ko.observableArray([]);
@@ -25,10 +26,12 @@ function BeerViewModel() {
 
     // 初期データの読み込み
     self.loadBeers = function () {
-        fetch('/api/beer')
+        //const currentUserId = localStorage.getItem('user_id');
+
+        fetch(`/api/beer`)
             .then(res => res.json())
             .then(data => {
-                const mapped = data.map(item => new Beer(item));
+                const mapped = data.map(item => new BeerModel(item));
                 self.beers(mapped);
             })
             .catch(err => console.error('error:', err));
@@ -58,16 +61,31 @@ function BeerViewModel() {
 
     // 削除処理（APIにDELETEを送る）
     self.deleteBeer = function (beer) {
+        if (!confirm(`You would like to delete ${beer.name()}?`)) {
+            return;
+        }
+    
         fetch(`/api/beer/${beer.id}`, {
             method: 'DELETE'
         })
-        .then(() => {
-            self.beers.remove(beer);
+        .then(res => res.json())
+        .then(data => {
+            if (data.success) {
+                self.beers.remove(beer);
+                alert('Success to delete!');
+            } else {
+                alert('Failed to delete...');
+                console.error(data.error);
+            }
         })
-        .catch(err => console.error('error:', err));
+        .catch(err => {
+            alert('An error occured...');
+            console.error(err);
+        });
     };
 
     self.loadBeers(); // 初期化時に呼ぶ
+
 }
 
-ko.applyBindings(new BeerViewModel());
+ko.applyBindings(new BeerListViewModel());
